+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0 - 2018-08-30
+
+- Add `target` option
+- Add `options` option
+- Improve some internals
+
 ## 0.1.0 - 2018-08-28
 
 - Initial release